add warning for repetitive entry

+++ website/ohara/views.py
@@ -3,6 +3,7 @@ from django.http import HttpResponse, HttpResponseRedirect
 from django.template import loader
 from django.urls import reverse
 from src.ohara import Ohara
+import os
 # Create your views here.
 
 def upload(request):
@@ -14,12 +15,30 @@ def text_handler(request):
     bibtex = request.POST['bibtex']
     lines = bibtex.split('\n')
     firstline = lines[0]
-    fname = firstline[firstline.find("/")+1:firstline.find(",")]
-    fpath = bibDir+fname+'.bib'
-    with open(fpath, 'w', encoding = 'utf-8') as f:
-        f.write(bibtex)
-    ohara.add_entry_from_file(fpath)
+    if firstline.find('/') >= 0:
+        fname = firstline[firstline.find("/")+1:firstline.find(",")]+'.bib'
+    else: 
+        fname = firstline[firstline.find("{")+1:firstline.find(",")]+'.bib'
+    fpath = bibDir+fname
+    file_exist = False
+    for dirName, subdirList, fileList in os.walk(ohara.rootDir):
+        if fname in fileList:
+            file_exist = True
+            break
+    if file_exist:
+        print(reverse('ohara:repeat_entry_warning', args = (fname,)))
+        return HttpResponseRedirect(reverse('ohara:repeat_entry_warning', args = (fname,)))
+    else:
+        with open(fpath, 'w', encoding = 'utf-8') as f:
+            f.write(bibtex)
+        ohara.add_entry_from_file(fpath)
     return HttpResponseRedirect(reverse('ohara:success'))
+
+def repeat_entry_warning(request, fname):
+    warning_msg = fname + " is already in Ohara!"
+    return render(request, 'ohara/warning.html', {
+            'warning_msg': warning_msg,
+        })
 
 def success(request):
     return render(request, 'ohara/success.html', {})
